refactor: expand utils module
