+++ scripts/plot_leading_order.py
@@ -7,7 +7,11 @@ from matplotlib import colors
 from photosurfactant.semi_analytic.leading_order import LeadingOrder
 from photosurfactant.parameters import Parameters, PlottingParameters
 from photosurfactant.semi_analytic.limits import HighIntensity, LowIntensity
-from photosurfactant.utils import leading_order_parser, parameter_parser, plot_parser
+from photosurfactant.utils.arg_parser import (
+    leading_order_parser,
+    parameter_parser,
+    plot_parser,
+)
 
 # TODO: Reduced figures at 3 magnitudes?
